Add Clash.Normalize.Transformations.MultiPrim
All transformations concerning primtives with multiple results now
live in a separate module. This is part of work to make the
normalization modules quicker to compile.
